Import GeometricOptimizers selectively, and move the optimizer code after the manifolds
GML did not precompile on Julia 1.10 at all:

    cannot assign a value to imported variable GeometricOptimizers.Manifold

`using GeometricOptimizers` was a blanket import, and GO exports about twenty
names GML defines itself -- `Manifold`, `StiefelManifold`, `GrassmannManifold`,
`SkewSymMatrix`, `SymmetricMatrix`, `LowerTriangular`, `UpperTriangular`,
`StiefelLieAlgHorMatrix`, `GrassmannLieAlgHorMatrix`, `Optimizer`,
`AdamOptimizerWithDecay`, `rgrad`, ... Julia 1.12's binding partitions tolerate
redefining an imported binding; 1.10 does not. Replaced with
`import GeometricOptimizers` plus an explicit `using GeometricOptimizers: ...`
list, extended with the names GML re-exports.

That exposed a silent correctness bug. `include("utils.jl")` runs at line 71 and
`include("manifolds/abstract_manifold.jl")` at line 142, so `Manifold` in

    _gml_rgrad(x::Manifold, dp) = rgrad(x, dp)

resolved to GeometricOptimizers.Manifold. GML's `StiefelManifold` does not
subtype that -- `GML.Manifold === GO.Manifold` is false, and GML's is
`<: AbstractMatrix` -- so the method never matched a GML manifold. The
`_gml_rgrad(x, dp) = dp` fallback caught it instead and the Riemannian
projection was skipped, passing the raw Euclidean gradient through. This is
issue #234's type split showing up somewhere new.

The optimizer machinery therefore moves out of `utils.jl` (lines 177-400) into
`src/optimizers/optimizer.jl`, included after the manifolds. The move is
byte-for-byte: `diff` against the old `utils.jl` section is empty, so the whole
change is the relocation plus the include. `_gml_rgrad` now dispatches to
`src/optimizers/optimizer.jl:16` for a `StiefelManifold`, as intended.

GML precompiles and loads on 1.10.11 and still on 1.12.6.
